+++ src/extension/models/constants.ts
@@ -1,7 +1,6 @@
 
 export enum Constants {
     // to notify, these all should be in caps
-    
     LANG_CODE = 'dothttp-vscode',
 
     // configs
@@ -17,7 +16,6 @@ export enum Constants {
     responseDirectory = "dothttp.conf.response.savedirectory",
     CONFIG_DOTHTTP_USE_STABLE = "dothttp.conf.useunstable",
     // view props
-
     // commands
     RUN_FILE_COMMAND = 'dothttp.command.run',
     GEN_CURL_FILE_COMMAND = 'dothttp.command.gencurl',
@@ -83,5 +81,6 @@ export enum Constants {
 
     dothttpVersion = "dothttp.version",
 
-    versionApi = "https://raw.githubusercontent.com/cedric05/dothttp-runner/VERSION/version.json"
+    versionApi = "https://raw.githubusercontent.com/cedric05/dothttp-runner/VERSION/version.json",
+    SECRET_POSTMAN_API_KEY = "POSTMAN_API_KEY"
 }